+++ Pod/SyneriseSDK.framework/Headers/SNRClientIdentityProvider.h
@@ -1,0 +1,25 @@
+//
+//  SNRClientIdentityProvider.h
+//  SyneriseSDK
+//
+//  Created by Synerise
+//  Copyright (c) 2020 Synerise. All rights reserved.
+//
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ * @enum SNRClientIdentityProvider
+ */
+
+typedef NS_ENUM(NSUInteger, SNRClientIdentityProvider) {
+    SNRClientIdentityProviderSynerise,
+    SNRClientIdentityProviderOAuth,
+    SNRClientIdentityProviderFacebook,
+    SNRClientIdentityProviderApple
+} NS_SWIFT_NAME(ClientIdentityProvider);
+
+NSString * SNR_ClientIdentityProviderToString(SNRClientIdentityProvider type);
+SNRClientIdentityProvider SNR_StringToClientIdentityProvider(NSString * _Nullable string);
+
+NS_ASSUME_NONNULL_END