+++ Pod/SyneriseSDK.framework/Headers/SNRContentWidgetBaseCustomAction.h
@@ -11,7 +11,7 @@ NS_ASSUME_NONNULL_BEGIN
 typedef NS_ENUM(NSUInteger, SNRContentWidgetBaseCustomActionPredefiniedActionType) {
     SNRContentWidgetBaseCustomActionPredefiniedActionTypeNone = 0,
     SNRContentWidgetBaseCustomActionPredefiniedActionTypeSendLikeEvent
-};
+} NS_SWIFT_NAME(ContentWidgetBaseCustomActionPredefiniedActionType);
 
 NS_SWIFT_NAME(ContentWidgetBaseCustomAction)
 @interface SNRContentWidgetBaseCustomAction : NSObject